+++ R/get_pvol_de.R
@@ -50,11 +50,12 @@ get_pvol_de <- function(radar, time, ..., call = rlang::caller_env()) {
   if (nrow(files_to_get) != 50) {
     cli::cli_abort(
       c(
-        x = "The server returned an unexpected number of files.",
-        i = "In many cases this because no data is available for the time requested. For Germany only the last three days are available.",
-        i = "Alternatively there is an unstable connection or temporal error in the German data provided."
+        "The server returned an unexpected number of files.",
+        "i" = "In many cases this because no data are available for the time
+               requested. For Germany only the last three days are available.
+               The error can also be caused by an unstable connection or
+               temporal error in the German data provided."
       ),
-
       class = "getRad_error_germany_unexpected_number_of_files",
       call = call
     )