+++ R/get_pvol_dk.R
@@ -3,7 +3,7 @@ get_pvol_dk <- function(radar, time, ..., call = rlang::caller_env()) {
     req <- httr2::request(
       getOption(
         "getRad.dk_url",
-        "https://dmigw.govcloud.dk/v1/radardata/download"
+        "https://opendataapi.dmi.dk/v1/radardata/download"
       )
     ) |>
       req_user_agent_getrad() |>
@@ -13,7 +13,6 @@ get_pvol_dk <- function(radar, time, ..., call = rlang::caller_env()) {
           "{radar}_{strftime(time,'%Y%m%d%H%M', tz='UTC')}.vol.h5"
         ))
       ) |>
-      httr2::req_url_query(`api-key` = get_secret("dk_api_key")) |>
       httr2::req_perform(path = "file.h5", error_call = call)
     pvol <- bioRad::read_pvolfile(req$body, ...)
   })