+++ tests/testthat/test-get_pvol_dk.R
@@ -1,9 +1,4 @@
 test_that("Pvol for Danish can be downloaded", {
-  withr::local_options(list("keyring_backend" = "env"))
-  skip_if(
-    inherits(try(get_secret("dk_api_key"), silent = TRUE), "try-error"),
-    message = "Because no key for Denmark is available in the testing environment"
-  )
   time <- lubridate::floor_date(
     as.POSIXct(Sys.time(), tz = "Europe/Helsinki") - lubridate::hours(10),
     "5 mins"