+++ tests/testthat/test-get_pvol_ro.R
@@ -7,10 +7,11 @@ test_that("Check if the available attributes changed", {
       xml2::xml_find_all("//a/@href") |>
       xml2::xml_text() |>
       tail(-1) |>
-      gsub(pattern = ".*0200", replacement = "") |>
+      gsub(pattern = "MED_[0-9]*00", replacement = "") |>
       unique() |>
-      gsub(pattern = ".hdf", replacement = ""),
-    c("KDP", "RhoHV", "V", "ZDR", "dBR", "dBZ", "Height")
+      gsub(pattern = ".hdf", replacement = "") |>
+      sort(),
+    c("KDP", "RhoHV", "V", "ZDR", "dBZ") |> sort()
   )
 })
 test_that("Pvol for Romania can be downloaded", {